src/services/backend/requirements.txt
asgi-lifespan==2.1.0
asgiref==3.7.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
defusedxml==0.7.1
Django==4.2.18
django-allauth==0.57.0
django-asgi-lifespan==0.4.0
django-cors-headers==4.2.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
h11==0.14.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==10.2.0
psycopg2-binary==2.9.6
pycparser==2.22
PyJWT==2.8.0
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
channels==4.1.0
channels-redis==4.1.0
daphne==4.1.0
redis==4.6.0